+++ evosax/strategies/pgpe.py
@@ -100,9 +100,9 @@ class PGPE(Strategy):
         """Update both mean and dim.-wise isotropic Gaussian scale."""
         # Reconstruct noise from last mean/std estimates
         noise = (x - state.mean) / state.sigma
-        noise_1 = noise[: int(self.popsize / 2)]
-        fit_1 = fitness[: int(self.popsize / 2)]
-        fit_2 = fitness[int(self.popsize / 2) :]
+        noise_1 = noise[::2]
+        fit_1 = fitness[::2]
+        fit_2 = fitness[1::2]
         elite_idx = jnp.minimum(fit_1, fit_2).argsort()[: self.elite_popsize]
 
         fitness_elite = jnp.concatenate([fit_1[elite_idx], fit_2[elite_idx]])